algofun-ai-service/requirements.txt
fastapi
uvicorn[standard]
google-genai
pandas
python-dotenv
